Only remove hot file if no longer hot

+++ src/Mix.js
@@ -67,9 +67,12 @@ class Mix {
      * Determine if Mix should activate hot reloading.
      */
     shouldHotReload() {
-        new File(path.join(Config.publicPath, 'hot')).delete();
+        const isHMR = this.isUsing('hmr');
+        if (!isHMR) {
+            new File(path.join(Config.publicPath, 'hot')).delete();
+        }
 
-        return this.isUsing('hmr');
+        return isHMR;
     }
